moved components to its folders
