+++ app/components/reviews/types.ts
@@ -1,0 +1,3 @@
+import { reviews } from "~/data/reviewsData.json";
+
+export type ReviewType = typeof reviews[number];
